refactor: move webpush to it own router
refactor to make `views/api.py` cleaner
