+++ lnbits/core/__init__.py
@@ -9,6 +9,7 @@ from .views.generic import generic_router, update_user_extension
 from .views.node_api import node_router, public_node_router, super_node_router
 from .views.public_api import public_router
 from .views.tinyurl_api import tinyurl_router
+from .views.webpush_api import webpush_router
 
 # backwards compatibility for extensions
 core_app = APIRouter(tags=["Core"])
@@ -24,3 +25,4 @@ def init_core_routers(app):
     app.include_router(public_node_router)
     app.include_router(admin_router)
     app.include_router(tinyurl_router)
+    app.include_router(webpush_router)
